requirements.txt
fastapi==0.115.12
Jinja2==3.1.6
mysql-connector-python==9.2.0
python-multipart==0.0.20
uvicorn==0.34.1